backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
PyJWT==2.10.1
Werkzeug==3.1.3
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
Flask-Mail==0.10.0
